Protect rm;!prepend;polish

+++ shared/scripts/copy.sh
@@ -26,7 +26,7 @@ done;
 
 "$temp_shared/scripts/copy_content_to_client.sh" "$setup_root" "$temp" "$temp_shared" "$temp_this" "$flags";
 
-rm -r "$temp/";
+rm -r "${temp:?}/";
 
 { printf "shopt -s expand_aliases\n\n"; cat ~/.bashrc.d/portable/aliases.bash; } > ~/vimfiles/bash_env.bash;
 { printf "setopt aliases\n\n"; cat ~/.bashrc.d/portable/aliases.bash; } > ~/vimfiles/bash_env.zsh;
